+++ post_isoseq_cluster/demux_isoseq_with_genome.py
@@ -2,7 +2,7 @@
 __author__ = 'etseng@pacb.com'
 
 """
-Demultiplex IsoSeq1/IsoSeq2 job output (with genome mapping)
+Demultiplex IsoSeq1/IsoSeq2/IsoSeq3 job output (with genome mapping)
 
 =================
 WITHOUT GENOME
@@ -17,12 +17,19 @@ INPUT: out_mapped.fastq, read_stat.txt, classify report (alternatively, job dire
 OUTPUT: CSV file containing associated FL count for each isoform:primer
 
 Read stat format:
+    (isoseq1 and isoseq2)
     id      length  is_fl   stat    pbid
     m54006_170729_232022/43123426/1712_71_CCS       1641    Y       unique  PB.3811.1
     m54006_170729_232022/26476826/32_1680_CCS       1648    Y       unique  PB.3811.1
     m54006_170729_232022/44958075/3189_67_CCS       3122    Y       unique  PB.3787.1
     m54006_170729_232022/9896496/3222_72_CCS        3150    Y       unique  PB.3787.1
-
+    (isoseq3)
+    id      length  is_fl   stat    pbid
+    m54043_180729_193105/31392528/ccs       5306    Y       unique  PB.6318.2
+    m54043_180729_193105/53346556/ccs       5233    Y       unique  PB.6318.2
+    m54043_180729_193105/66716361/ccs       5030    Y       unique  PB.8713.2
+    m54043_180729_193105/6488380/ccs        5001    Y       unique  PB.8713.2
+    
 Classify report format:
    (isoseq1 and 2)
     id,strand,fiveseen,polyAseen,threeseen,fiveend,polyAend,threeend,primer,chimera
@@ -38,12 +45,13 @@ Classify report format:
 """
 
 import os, re, sys
-from csv import DictReader
+from csv import DictReader, DictWriter
 from collections import defaultdict, Counter
 from Bio import SeqIO
 
 hq1_id_rex = re.compile('i\d+_HQ_\S+\|(\S+)\/f\d+p\d+\/\d+')
 hq2_id_rex = re.compile('HQ_\S+\|(\S+)\/f\d+p\d+\/\d+')
+hq3_id_rex = re.compile('[\S_]+(transcript/\d+)')
 
 def link_files(src_dir, out_dir='./'):
     """
@@ -56,11 +64,16 @@ def link_files(src_dir, out_dir='./'):
     # location for mapped fastq in IsoSeq2
     mapped_fastq2 = os.path.join(os.path.abspath(src_dir), 'tasks', 'pbtranscript2tools.tasks.post_mapping_to_genome-0', 'output_mapped.fastq')
     mapped_gff2 = os.path.join(os.path.abspath(src_dir), 'tasks', 'pbtranscript2tools.tasks.post_mapping_to_genome-0', 'output_mapped.gff')
+    # location for mapped fastq in IsoSeq3
+    mapped_fastq3 = os.path.join(os.path.abspath(src_dir), 'tasks', 'isocollapse.tasks.collapse_mapped_isoforms-0', 'out.fastq')
+    mapped_gff3 = os.path.join(os.path.abspath(src_dir), 'tasks', 'isocollapse.tasks.collapse_mapped_isoforms-0', 'out.gff')
     # location for read_stat.txt in IsoSeq1 IsoSeq2
     read_stat = os.path.join(os.path.abspath(src_dir), 'tasks', 'pbtranscript.tasks.post_mapping_to_genome-0', 'output_mapped.no5merge.collapsed.read_stat.txt')
     read_stat2 = os.path.join(os.path.abspath(src_dir), 'tasks', 'pbtranscript2tools.tasks.post_mapping_to_genome-0', 'output_mapped.no5merge.collapsed.read_stat.txt')
+    read_stat3 = os.path.join(os.path.abspath(src_dir), 'tasks', 'isocollapse.tasks.collapse_mapped_isoforms-0', 'out.no5merge.collapsed.read_stat.txt')
     # location for classify report in IsoSeq1 and 2
     primer_csv = os.path.join(os.path.abspath(src_dir), 'tasks', 'pbcoretools.tasks.gather_csv-1', 'file.csv')
+    lima_report = os.path.join(os.path.abspath(src_dir), 'tasks', 'barcoding.tasks.lima-0', 'lima_output.lima.report')
 
     if os.path.exists(mapped_fastq):
         print >> sys.stderr, "Detecting IsoSeq1 task directories..."
@@ -69,13 +82,23 @@ def link_files(src_dir, out_dir='./'):
         os.symlink(read_stat, os.path.join(out_dir, 'mapped.read_stat.txt'))
         os.symlink(primer_csv, os.path.join(out_dir, 'classify_report.csv'))
         isoseq_version = '1'
-    else:
+    elif os.path.exists(mapped_fastq2):
         print >> sys.stderr, "Detecting IsoSeq2 task directories..."
         os.symlink(mapped_fastq2, os.path.join(out_dir, 'mapped.fastq'))
         os.symlink(mapped_gff2, os.path.join(out_dir, 'mapped.gff'))
         os.symlink(read_stat2, os.path.join(out_dir, 'mapped.read_stat.txt'))
         os.symlink(primer_csv, os.path.join(out_dir, 'classify_report.csv'))
         isoseq_version = '2'
+    elif os.path.exists(mapped_fastq3):
+        print >> sys.stderr, "Detecting IsoSeq3 task directories..."
+        os.symlink(mapped_fastq3, os.path.join(out_dir, 'mapped.fastq'))
+        os.symlink(mapped_gff3, os.path.join(out_dir, 'mapped.gff'))
+        os.symlink(read_stat3, os.path.join(out_dir, 'mapped.read_stat.txt'))
+        make_classify_csv_from_lima_report(lima_report, os.path.join(out_dir, 'classify_report.csv'))
+        isoseq_version = '3'
+    else:
+        print >> sys.stderr, "Cannot find HQ FASTQ in job directory! Does not look like Iso-Seq1, 2, or 3 jobs!"
+        sys.exit(-1)
     return out_dir, 'mapped.fastq', 'mapped.read_stat.txt', 'classify_report.csv', isoseq_version
 
 def read_read_stat(read_stat, classify_info):
@@ -108,11 +131,26 @@ def read_classify_csv(classify_csv):
         info[r['id']] = p
     return primer_list, info
 
+def make_classify_csv_from_lima_report(report_filename, output_filename):
+
+    h = open(output_filename, 'w')
+    fout = DictWriter(h, fieldnames=['id', 'primer_index', 'primer'], delimiter=',')
+    fout.writeheader()
+    for r in DictReader(open(report_filename), delimiter='\t'):
+        assert r['IdxLowestNamed'].endswith('_5p') or r['IdxLowestNamed'].endswith('_3p')
+        assert r['IdxHighestNamed'].endswith('_5p') or r['IdxHighestNamed'].endswith('_3p')
+        if r['IdxLowestNamed'].endswith('_5p') and r['IdxHighestNamed'].endswith('_3p'):
+            newrec = {'id': r['ZMW']+'/ccs',
+                    'primer_index': "{0}--{1}".format(r['IdxLowest'], r['IdxHighest']),
+                    'primer': "{0}--{1}".format(r['IdxLowestNamed'], r['IdxHighestNamed'])}
+        fout.writerow(newrec)
+    h.close()
+
 
 def main(job_dir=None, mapped_fastq=None, read_stat=None, classify_csv=None, output_filename=sys.stdout, primer_names=None):
     if job_dir is not None:
         out_dir_ignore, mapped_fastq, read_stat, classify_csv, isoseq_version = link_files(job_dir)
-        assert isoseq_version in ('1', '2')
+        assert isoseq_version in ('1', '2', '3')
     else:
         assert os.path.exists(mapped_fastq)
         assert os.path.exists(read_stat)
